+++ src/api/PersonsApi.js
@@ -17,6 +17,7 @@ import AddFollowerToPersonResponse from '../model/AddFollowerToPersonResponse';
 import AddPersonFollowerRequest from '../model/AddPersonFollowerRequest';
 import AddPersonPictureResponse from '../model/AddPersonPictureResponse';
 import AddPersonResponse from '../model/AddPersonResponse';
+import ChangelogResponse from '../model/ChangelogResponse';
 import DeletePersonResponse from '../model/DeletePersonResponse';
 import DeletePersonsInBulkResponse from '../model/DeletePersonsInBulkResponse';
 import FailResponse from '../model/FailResponse';
@@ -709,6 +710,82 @@ export default class PersonsApi {
      */
     getPersonActivities(id, opts) {
       return this.getPersonActivitiesWithHttpInfo(id, opts)
+        .then(function(response_and_data) {
+          return response_and_data;
+        });
+    }
+
+
+    /**
+     * List updates about person field values
+     * Lists updates about field values of a person.
+     * @param {Number} id The ID of the person
+     * @param {Object} opts Optional parameters
+     * @param {String} opts.cursor For pagination, the marker (an opaque string value) representing the first item on the next page
+     * @param {Number} opts.limit Items shown per page
+     * @return {Promise} a {@link https://www.promisejs.org/|Promise}, with an object containing data of type {@link module:model/ChangelogResponse} and HTTP response
+     */
+    getPersonChangelogWithHttpInfo(id, opts) {
+      opts = opts || {};
+      let postBody = null;
+
+      // verify the required parameter 'id' is set
+      if (id === undefined || id === null) {
+        throw new Error("Missing the required parameter 'id' when calling getPersonChangelog");
+      }
+
+
+
+      let pathParams = {
+        'id': id,
+      };
+      let queryParams = {
+        'cursor': opts['cursor'] === undefined ? opts['cursor'] : opts['cursor'],
+        'limit': opts['limit'] === undefined ? opts['limit'] : opts['limit'],
+      };
+      let headerParams = {
+      };
+      let formParams = {
+      };
+
+      let formParamArray = [
+      ];
+
+      let contentTypes = [];
+      const isURLEncoded = contentTypes.includes('application/x-www-form-urlencoded');
+      const isJSON = contentTypes.includes('application/json');
+
+      if (isJSON) {
+        postBody = { ...postBody, ...opts };
+      } else if (isURLEncoded) {
+        for (let key in opts) {
+          if (opts.hasOwnProperty(key) && !formParamArray.includes(key)) {
+            formParams[key] = opts[key];
+          }
+        }
+      }
+
+      let authNames = ['api_key', 'oauth2', ];
+      let accepts = ['application/json', ];
+      let returnType = ChangelogResponse;
+      return this.apiClient.callApi(
+        '/persons/{id}/changelog', 'GET',
+        pathParams, queryParams, headerParams, formParams, postBody,
+        authNames, contentTypes, accepts, returnType, null
+      );
+    }
+
+    /**
+     * List updates about person field values
+     * Lists updates about field values of a person.
+     * @param {Number} id The ID of the person
+     * @param {Object} opts Optional parameters
+     * @param {String} opts.cursor For pagination, the marker (an opaque string value) representing the first item on the next page
+     * @param {Number} opts.limit Items shown per page
+     * @return {Promise} a {@link https://www.promisejs.org/|Promise}, with data of type {@link module:model/ChangelogResponse}
+     */
+    getPersonChangelog(id, opts) {
+      return this.getPersonChangelogWithHttpInfo(id, opts)
         .then(function(response_and_data) {
           return response_and_data;
         });