+++ src/model/ActivityTypeObjectResponse.js
@@ -89,12 +89,12 @@ class ActivityTypeObjectResponse {
                 delete data['is_custom_flag'];
             }
             if (data.hasOwnProperty('add_time')) {
-                obj['add_time'] = ApiClient.convertToType(data['add_time'], 'String');
+                obj['add_time'] = ApiClient.convertToType(data['add_time'], 'Date');
 
                 delete data['add_time'];
             }
             if (data.hasOwnProperty('update_time')) {
-                obj['update_time'] = ApiClient.convertToType(data['update_time'], 'String');
+                obj['update_time'] = ApiClient.convertToType(data['update_time'], 'Date');
 
                 delete data['update_time'];
             }
@@ -159,13 +159,13 @@ ActivityTypeObjectResponse.prototype['is_custom_flag'] = undefined;
 
 /**
  * The creation time of the activity type
- * @member {String} add_time
+ * @member {Date} add_time
  */
 ActivityTypeObjectResponse.prototype['add_time'] = undefined;
 
 /**
  * The update time of the activity type
- * @member {String} update_time
+ * @member {Date} update_time
  */
 ActivityTypeObjectResponse.prototype['update_time'] = undefined;